refactor: invert cast logic to store integers in DB, work with decimals in app
BREAKING CHANGE: This completely changes how the cast works

**OLD behavior (incorrect):**
- DB: DECIMAL (19.99) → App: INTEGER (1999)

**NEW behavior (correct):**
- DB: INTEGER (1999) → App: DECIMAL (19.99)

**What changed:**
- Base100 cast logic inverted (get/set methods swapped)
- Database migration now uses INTEGER instead of DECIMAL
- All tests updated to reflect new behavior
- README completely rewritten with correct examples
- Users work with natural decimal values (19.99)
- Database stores efficient integers (1999 cents)

**Migration required:**
Users need to change their migrations from:
```php
$table->decimal('price', 10, 2);  // OLD
```
to:
```php
$table->integer('price');  // NEW
```

This is the standard approach used by Money libraries.

+++ tests/database/migrations/create_test_models_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     {
         Schema::create('test_models', function (Blueprint $table) {
             $table->id();
-            $table->decimal('price', 10, 2)->default(0);
-            $table->decimal('cost', 10, 2)->default(0);
-            $table->decimal('tax', 10, 2)->default(0);
+            $table->integer('price')->default(0);  // Stores cents: 1999 = $19.99
+            $table->integer('cost')->default(0);   // Stores cents: 1500 = $15.00
+            $table->integer('tax')->default(0);    // Stores cents: 250 = $2.50
             $table->timestamps();
         });
     }
